Move composeMessage function to rc-utils
