+++ packages/rocketchat-lib/server/functions/composeMessageObjectWithUser_import.js
@@ -1,0 +1,3 @@
+import { composeMessageObjectWithUser } from 'meteor/rocketchat:utils';
+
+RocketChat.composeMessageObjectWithUser = composeMessageObjectWithUser;
