+++ packages/rocketchat-utils/server/index.js
@@ -14,3 +14,4 @@ export { getAvatarColor } from '../lib/getAvatarColor';
 export { getURL } from '../lib/getURL';
 export { getValidRoomName } from '../lib/getValidRoomName';
 export { placeholders } from '../lib/placeholders';
+export { composeMessageObjectWithUser } from './lib/composeMessageObjectWithUser';
